Run a full set of Latent Dirichlet Allocations and Time Series models

Conduct a complete LDATS analysis (Christensen et al. 2018), including running a suite of Latent Dirichlet Allocation (LDA) models (Blei et al. 2003, Grun and Hornik 2011) via LDA_set, selecting LDA model(s) via select_LDA, running a complete set of Bayesian Time Series (TS) models (Western and Kleykamp 2004) via TS_on_LDA on the chosen LDA model(s), and selecting the best TS model via select_TS.

check_LDA_TS_inputs checks that the inputs to LDA_TS are of proper classes for a full analysis.

LDA_TS(document_term_table, document_covariate_table, topics = 2,
  nseeds = 1, formulas = ~1, nchangepoints = 0,
  weights = document_weights(document_term_table),
  control = LDA_TS_controls_list())

check_LDA_TS_inputs(document_term_table, document_covariate_table,
  topics = 2, nseeds = 1, formulas = ~1, nchangepoints = 0,
  weights = document_weights(document_term_table),
  control = LDA_TS_controls_list())

Arguments

document_term_table
Table of observation count data (rows: documents (\(M\)), columns: terms (\(V\))). May be a matrix or data.frame but must be conformable to a matrix of integers (as verified by check_document_term_table). This table is a document-level summary of the data noted as w (the word-level topic identity) in the math description.

document_covariate_table
Document covariate table (rows: documents (\(M\)), columns: time index and covariate options). Every model needs a covariate to describe the time value for each document (in whatever relevant units), whose name in the table is input in control (the timename entry), that dictates the application of the change points. In addition, all covariates named within specific models in formula must be included. Must be a conformable to a data table, as verified by check_document_covariate_table.

topics
Vector of the number of topics to evaluate in the LDAs. Must be integer values.

nseeds
integer number of seeds (replicate starts) to use for each value of topics in the LDAs.

formulas
Vector of formula(s) for the continuous (non-changepoint) component of the time series models. Any predictor variable included in a formula must also be a column in the document_covariate_table. Each element (formula) in the vector is evaluated for each number of change points and each LDA model.

nchangepoints
Vector of integers corresponding to the number of change points to include in the time series models. 0 is a valid input corresponding to no change points (i.e., a singular time series model), and the current implementation can reasonably include up to 6 change points. Each element in the vector is the number of change points (Pm2 for model m2 in the math description) used to segement the data for each formula (entry in formulas) component of the TS model, for each selected LDA model.

weights
Optional class numeric vector of weights for each document. Corresponds to the vector \(u\) in the math description. Defaults to NULL, translating to an equal weight for each document. When using multinom_TS in a standard LDATS analysis, it is advisable to weight the documents by their total size, as the result of LDA is a matrix of proportions ( Γ ), which does not account for size differences among documents. For most models, a scaling of the weights (so that the average is 1) is most appropriate, and this is accomplished using document_weights

control
Class LDA_TS_controls list (as generated by LDA_TS_controls_list) that contains LDA_controls (LDA_controls_list), TS_controls (TS_controls_list), and the single top-level control quiet.

Value

LDATS returns a class LDA_TS list object including all fitted LDA and TS models and selected models specifically as elements "LDA models" (from LDA_set), "Selected LDA model" (from select_LDA), "TS models" (from TS_on_LDA), and "Selected TS model" (from select_TS).

References

Blei, D. M., A. Y. Ng, and M. I. Jordan. 2003. Latent Dirichlet Allocation. Journal of Machine Learning Research 3:993-1022. link.

Christensen, E., D. J. Harris, and S. K. M. Ernest. 2018. Long-term community change through multiple rapid transitions in a desert rodent community. Ecology 99:1523-1529. link.

Grun B. and K. Hornik. 2011. topicmodels: An R Package for Fitting Topic Models. Journal of Statistical Software 40:13. link.

Western, B. and M. Kleykamp. 2004. A Bayesian change point model for historical time series analysis. Political Analysis 12:354-374. link.

Examples

# NOT RUN {
  data(rodents)
  document_term_table <- rodents$document_term_table
  document_covariate_table <- rodents$document_covariate_table

  mod <- LDA_TS(document_term_table, document_covariate_table,
                topics = 2, nseeds = 1, formulas = ~ 1, nchangepoints = 1,
                weights = document_weights(document_term_table),
                control = LDA_TS_controls_list())
# }