Create the controls list for the Time Series model

This function provides a simple creation and definition of a list used to control the time series model fit occurring within TS.

TS_controls_list(memoise = TRUE, response = "gamma",
  timename = "time", lambda = 0, measurer = AIC, selector = min,
  ntemps = 6, penultimate_temp = 2^6, ultimate_temp = 1e+10, q = 0,
  nit = 10000, magnitude = 12, quiet = FALSE, burnin = 0,
  thin_frac = 1, summary_prob = 0.95, seed = NULL)

Arguments

memoise
logical indicator of whether the multinomial functions should be memoised (via memoise). Memoisation happens to both multinom_TS and multinom_TS_chunk.

response
character element indicating the response variable used in the time series.

timename
character element indicating the time variable used in the time series. Defaults to "time".

lambda
numeric "weight" decay term used to set the prior on the regressors within each chunk-level model. Defaults to 0, corresponding to a fully vague prior.

measurer, selector
Function names for use in evaluation of the TS models. measurer is used to create a value for each model and selector operates on the values to choose the model.

ntemps
integer number of temperatures (chains) to use in the ptMCMC algorithm.

penultimate_temp
Penultimate temperature in the ptMCMC sequence.

ultimate_temp
Ultimate temperature in the ptMCMC sequence.

q
Exponent controlling the ptMCMC temperature sequence from the focal chain (reference with temperature = 1) to the penultimate chain. 0 (default) implies a geometric sequence. 1 implies squaring before exponentiating.

nit
integer number of iterations (steps) used in the ptMCMC algorithm.

magnitude
Average magnitude (defining a geometric distribution) for the proposed step size in the ptMCMC algorithm.

quiet
logical indicator of whether the model should run quietly (if FALSE, a progress bar and notifications are printed).

burnin
integer number of iterations to remove from the beginning of the ptMCMC algorithm.

thin_frac
Fraction of iterations to retain, must be \((0, 1]\), and the default value of 1 represents no thinning.

summary_prob
Probability used for summarizing the posterior distributions (via the highest posterior density interval, see HPDinterval).

seed
Input to set.seed for replication purposes.

Value

List of class TS_controls, with named elements corresponding to the arguments.