Prepare the model-specific data to be used in the TS analysis of LDA output

Append the estimated topic proportions from a fitted LDA model to the document covariate table to create the data structure needed for TS.

prep_TS_data(document_covariate_table, LDA_models, mods, i)

Arguments

document_covariate_table
Document covariate table (rows: documents, columns: time index and covariate options). Every model needs a covariate to describe the time value for each document (in whatever units and whose name in the table is input in control$timename) that dictates the application of the change points. In addition, all covariates named within specific models in formula must be included. Must be a conformable to a data table, as verified by check_document_covariate_table.

LDA_models
List of LDA models (class LDA_set, produced by LDA_set) or a singular LDA model (class LDA, produced by LDA).

mods
The data.table created by expand_TS that contains each of the models (defined by the LDA model to use and the and formula number of changepoints for the TS model). Indexded here by i.

i
integer index referencing the row in mods to use.

Value

Class data.frame object including [1] the time variable (indicated in control), [2] the predictor variables (required by formula) and [3], the multinomial response variable (indicated in formula), ready for input into TS.