Initialize and update the changepoint matrix used in the ptMCMC algorithm

Each of the chains is initialized by prep_cpts using a draw from the available times (i.e. assuming a uniform prior), the best fit (by likelihood) draw is put in the focal chain with each subsequently worse fit placed into the subsequently hotter chain. update_cpts updates the changepoints after every iteration in the ptMCMC algorithm.

prep_cpts(data, formula, nchangepoints, weights,
  control = TS_controls_list())

update_cpts(cpts, swaps)

Arguments

data
Class data.frame object including [1] the time variable (indicated in control), [2] the predictor variables (required by formula) and [3], the multinomial response variable (indicated in formula).

formula
formula describing the continuous change. Any predictor variable included must also be a column in the data. Any (multinomial) response variable must also be a set of columns in data.

nchangepoints
Integer corresponding to the number of change points to include in the model. 0 is a valid input (corresponding to no change points, so a singular time series model), and the current implementation can reasonably include up to 6 change points. The number of change points is used to dictate the segementation of the data for each continuous model and each LDA model.

weights
Optional class numeric vector of weights for each document. Corresponds to the vector \(v\) in the math description.

control
Class TS_controls list, holding control parameters for the Time Series model including the parallel tempering Markov Chain Monte Carlo (ptMCMC) controls, generated by TS_controls_list.

cpts
The existing matrix of changepoints.

swaps
Chain configuration after among-temperature swaps.

Value

List of [1] matrix of change points (rows) for each temperature (columns) and [2] vector of log-likelihood values for each of the chains.